requirements.txt
botbuilder-core
botbuilder-schema
azure-identity
azure-core
langchain-openai  
pyodbc
aioodbc
nest_asyncio
tabulate
langchain
botbuilder-integration-aiohttp
langchain-cli
botbuilder-dialogs
pyjwt
cryptography
faiss-cpu
regex
tiktoken
azure-storage-blob
langchain_community
msal
pandas